+++ lib.php
@@ -42,7 +42,7 @@ function tool_courserating_before_http_headers() {
         $PAGE->requires->js_call_amd(
             'tool_courserating/rating',
             'init',
-            [context_system::instance()->id, "{$branch}" < "400"]
+            [context_system::instance()->id]
         );
         if (\tool_courserating\helper::is_course_edit_page()) {
             $field = \tool_courserating\helper::get_course_rating_field();